requirements.txt
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-cncf-kubernetes==7.11.0
pyspark==3.4.1
hdfs==2.7.0
pandas==2.1.1
sqlalchemy==1.4.50
